api improvements
