+++ routes/api.yml
@@ -14,6 +14,7 @@ tags:
   - name: languages
   - name: misc
   - name: suggestions
+  - name: notes
 
 paths:
   /cheevos:
@@ -140,6 +141,111 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/Language'
+  
+  /notes:
+    get:
+      operationId: getUserNotes
+      tags: [notes]
+      summary: Get all current golfer's notes
+      security:
+        - cookieAuth: []
+      responses:
+        '200':
+          description: All golfer's notes
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/Note'
+        '403':
+          description: Notes access requires authentication
+  
+  /notes/{hole}/{lang}:
+    get:
+      operationId: getUserNote
+      parameters:
+        - name: hole
+          description: Hole ID
+          in: path
+          required: true
+          schema:
+            type: string
+        - name: lang
+          description: Language ID
+          in: path
+          required: true
+          schema:
+            type: string
+      tags: [notes]
+      summary: Get current golfer's note for given hole, lang pair
+      security:
+        - cookieAuth: []
+      responses:
+        '200':
+          description: A golfer's note
+          content:
+            application/json:
+              schema:
+                type: string
+        '403':
+          description: Notes access requires authentication
+        '404':
+          description: Note does not exist
+    put:
+      operationId: upsertUserNote
+      parameters:
+        - name: hole
+          description: Hole ID
+          in: path
+          required: true
+          schema:
+            type: string
+        - name: lang
+          description: Language ID
+          in: path
+          required: true
+          schema:
+            type: string
+      tags: [notes]
+      summary: Creates or updates current golfer's note for given hole, lang pair
+      security:
+        - cookieAuth: []
+      responses:
+        '204':
+          description: Success
+        '400':
+          description: Invalid hole, lang, or note length
+        '402':
+          description: Notes are a sponsor-only feature
+        '403':
+          description: Notes access requires authentication
+    delete:
+      operationId: deleteUserNote
+      parameters:
+        - name: hole
+          description: Hole ID
+          in: path
+          required: true
+          schema:
+            type: string
+        - name: lang
+          description: Language ID
+          in: path
+          required: true
+          schema:
+            type: string
+      tags: [notes]
+      summary: Deletes current golfer's note for given hole, lang pair, if it exists
+      security:
+        - cookieAuth: []
+      responses:
+        '204':
+          description: Success
+        '400':
+          description: Invalid hole or lang
+        '403':
+          description: Notes access requires authentication
 
   /suggestions/golfers:
     get:
@@ -211,6 +317,11 @@ paths:
                 $ref: '#/components/schemas/WikiPage'
 
 components:
+  securitySchemes:
+    cookieAuth:
+      type: apiKey
+      in: cookie
+      name: __Host-session
   schemas:
     Achievement:
       type: object
@@ -285,7 +396,17 @@ components:
           type: string
         website:
           type: string
-          format: uri
+          format: uri  
+  
+    Note:
+      type: object
+      properties:
+        hole:
+          type: string
+        lang:
+          type: string
+        note:
+          type: string
 
     SolutionsLog:
       type: object
