+++ routes/legacy_api.go
@@ -1,7 +1,9 @@
 package routes
 
 import (
+	_ "embed"
 	"encoding/json"
+	"fmt"
 	"net/http"
 	"slices"
 	"time"
@@ -13,6 +15,55 @@ import (
 	"github.com/code-golf/code-golf/session"
 	"github.com/lib/pq"
 )
+
+// GET /scores/{hole}/{lang}/all
+func scoresAllGET(w http.ResponseWriter, r *http.Request) {
+	var json []byte
+
+	if err := session.Database(r).QueryRow(
+		`WITH solution_lengths AS (
+		    SELECT hole,
+		           lang,
+		           scoring,
+		           login,
+		           chars,
+		           bytes,
+		           submitted
+		      FROM solutions
+		      JOIN users ON user_id = users.id
+		     WHERE NOT failing
+		       AND $1 IN ('all-holes', hole::text)
+		       AND $2 IN ('all-langs', lang::text)
+		  ORDER BY submitted DESC
+		) SELECT COALESCE(JSON_AGG(solution_lengths), '[]') FROM solution_lengths`,
+		param(r, "hole"),
+		param(r, "lang"),
+	).Scan(&json); err != nil {
+		panic(err)
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.Write(json)
+}
+
+// GET /scores/{hole}/{lang}/{scoring}
+func scoresGET(w http.ResponseWriter, r *http.Request) {
+	holeID := param(r, "hole")
+	langID := param(r, "lang")
+	scoring := param(r, "scoring")
+
+	if holeID == "all-holes" {
+		holeID = "all"
+	}
+	if langID == "all-holes" {
+		langID = "all"
+	}
+	if scoring == "" {
+		scoring = "bytes"
+	}
+
+	http.Redirect(w, r, "/rankings/holes/"+holeID+"/"+langID+"/"+scoring, http.StatusPermanentRedirect)
+}
 
 // POST /solution
 func solutionPOST(w http.ResponseWriter, r *http.Request) {
@@ -265,4 +316,149 @@ func solutionPOST(w http.ResponseWriter, r *http.Request) {
 	if err := enc.Encode(&out); err != nil {
 		panic(err)
 	}
+}
+
+// GET /mini-rankings/{hole}/{lang}/{scoring:bytes|chars}/{view:top|me|following}
+func apiMiniRankingsGET(w http.ResponseWriter, r *http.Request) {
+	limit := 7
+	if r.FormValue("long") == "1" {
+		limit = 99
+	}
+
+	var (
+		holeID  = param(r, "hole")
+		langID  = param(r, "lang")
+		scoring = param(r, "scoring")
+		view    = param(r, "view")
+	)
+
+	// No need to check scoring & view, they're covered by chi.
+	hole := config.AllHoleByID[holeID]
+	lang := config.AllLangByID[langID]
+	if hole == nil || lang == nil {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	} else if hole.Experiment != 0 || lang.Experiment != 0 {
+		w.Write([]byte("[]"))
+		return
+	}
+
+	otherScoring := "bytes"
+	if scoring == "bytes" {
+		otherScoring = "chars"
+	}
+
+	var followLimit, userID int
+	if golfer := session.Golfer(r); golfer != nil {
+		followLimit = golfer.FollowLimit()
+		userID = golfer.ID
+	}
+
+	type entry struct {
+		Bytes      *int `json:"bytes"`
+		BytesChars *int `json:"bytes_chars"`
+		Chars      *int `json:"chars"`
+		CharsBytes *int `json:"chars_bytes"`
+		Golfer     struct {
+			ID   int    `json:"id"`
+			Name string `json:"name"`
+		} `json:"golfer"`
+		Me   bool `json:"me"`
+		Rank int  `json:"rank"`
+	}
+
+	sqlWhere, sqlLimit := "true", "$6"
+	switch view {
+	case "me":
+		sqlWhere = "row > COALESCE((SELECT row FROM ranks WHERE me), 0) - $6"
+		sqlLimit = "$6 * 2"
+	case "following":
+		sqlWhere = fmt.Sprint("user_id = ANY(following($1, ", followLimit, "))")
+	}
+
+	// We don't use the rankings view as we want instant updates upon solution
+	// submit, therefore we skip scoring to keep it fast.
+	rows, err := session.Database(r).Query(
+		`WITH ranks AS (
+		    SELECT ROW_NUMBER() OVER (ORDER BY `+scoring+`, submitted) row,
+		           RANK()       OVER (ORDER BY `+scoring+`),
+		           user_id,
+		           `+scoring+`,
+		           `+otherScoring+` `+scoring+`_`+otherScoring+`,
+		           user_id = $1 me
+		      FROM solutions
+		     WHERE hole = $2
+		       AND lang = $3
+		       AND scoring = $4
+		       AND NOT failing
+		), other_scoring AS (
+		    SELECT user_id,
+		           `+otherScoring+`,
+		           `+scoring+` `+otherScoring+`_`+scoring+`
+		      FROM solutions
+		     WHERE hole = $2
+		       AND lang = $3
+		       AND scoring = $5
+		       AND NOT failing
+		)   SELECT bytes, bytes_chars, chars, chars_bytes, id, login, me, rank
+		      FROM ranks
+		      JOIN users ON id = user_id
+		 LEFT JOIN other_scoring USING(user_id)
+		     WHERE `+sqlWhere+`
+		  ORDER BY row
+		     LIMIT `+sqlLimit,
+		userID,
+		holeID,
+		langID,
+		scoring,
+		otherScoring,
+		limit,
+	)
+	if err != nil {
+		panic(err)
+	}
+	defer rows.Close()
+
+	entries := make([]entry, 0, limit)
+	for rows.Next() {
+		var e entry
+
+		if err := rows.Scan(
+			&e.Bytes,
+			&e.BytesChars,
+			&e.Chars,
+			&e.CharsBytes,
+			&e.Golfer.ID,
+			&e.Golfer.Name,
+			&e.Me,
+			&e.Rank,
+		); err != nil {
+			panic(err)
+		}
+
+		entries = append(entries, e)
+	}
+
+	if err := rows.Err(); err != nil {
+		panic(err)
+	}
+
+	// Trim the rows to limit with "me" as close to the middle as possible.
+	// TODO It would simplify everything if we could fold this into the SQL.
+	length := len(entries)
+	if view == "me" && length > limit {
+		me := slices.IndexFunc(entries, func(e entry) bool { return e.Me })
+		// Before: me entries, then "me" entry, then len(entries)-me-1 entries
+		// 	with me <= limit; len(entries) <= 2*limit; len(entries)-me-1 <= limit-1
+		if me <= limit/2 {
+			entries = entries[:limit]
+		} else if me >= length-(limit+1)/2 {
+			// Impossible case?
+			entries = entries[length-limit:]
+		} else {
+			entries = entries[me-limit/2 : me+(limit+1)/2]
+		}
+	}
+
+	encodeJSON(w, entries)
 }
